+++ frontend/app/[locale]/setup/agent/components/AgentSetupOrchestrator.tsx
