+++ frontend/app/[locale]/setup/agent/components/DebugConfig.tsx
