+++ frontend/app/[locale]/setup/agent/components/McpConfigModal.tsx
