+++ frontend/app/[locale]/setup/agent/components/PromptManager.tsx
