+++ frontend/app/[locale]/setup/agent/components/agent/AgentConfigModal.tsx
