+++ frontend/app/[locale]/setup/agent/components/agent/CollaborativeAgentDisplay.tsx
