+++ frontend/app/[locale]/setup/agent/components/tool/ToolConfigModal.tsx
