+++ frontend/app/[locale]/setup/agent/components/tool/ToolPool.tsx
