+++ frontend/app/[locale]/setup/agent/config.tsx
