+++ frontend/app/[locale]/setup/knowledge/components/document/DocumentList.tsx
