+++ frontend/app/[locale]/setup/knowledge/components/document/DocumentStatus.tsx
