+++ frontend/app/[locale]/setup/knowledge/components/knowledge/KnowledgeBaseList.tsx
