+++ frontend/app/[locale]/setup/knowledge/components/upload/UploadArea.tsx
