+++ frontend/app/[locale]/setup/knowledge/components/upload/UploadAreaUI.tsx
