+++ frontend/app/[locale]/setup/knowledge/config.tsx
