+++ frontend/app/[locale]/setup/model/components/appConfig.tsx
