+++ frontend/app/[locale]/setup/model/components/model/ModelDeleteDialog.tsx
