+++ frontend/app/[locale]/setup/model/components/model/ModelListCard.tsx
