+++ frontend/app/[locale]/setup/model/components/modelConfig.tsx
