+++ frontend/app/[locale]/setup/model/config.tsx
